requirements.txt
mcp>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.25.0
python-dotenv>=1.0.0
PyPortfolioOpt==1.5.6